+++ R/glance_custom.R
@@ -39,8 +39,19 @@ glance_custom_internal.fixest <- function(x, vcov_type = NULL, ...) {
   for (n in x$fixef_vars) {
     out[[paste('FE:', n)]] <- 'X'
   }
+  # if (vcov=="robust") {
+  #   vcov = NULL
+  # }
   if (is.null(vcov_type) || !vcov_type %in% c("vector", "matrix", "function")) {
-    out[['vcov.type']] <- attr(fixest::coeftable(x), "type")
+    fvcov_type <- attr(fixest::coeftable(x), "type")
+    if (utils::packageVersion("fixest") >= "0.10.0") {
+      if (grepl("^Clustered", fvcov_type)) fvcov_type = gsub("\\)$", "", fvcov_type)
+      fvcov_type <- gsub("^Clustered \\(", "by: ", fvcov_type)
+    } else {
+      fvcov_type <- gsub("^Two-way|^Three-way|^Four-way", "", fvcov_type)
+      fvcov_type <- gsub("^ \\(", "by: ", gsub("\\)$", "", fvcov_type))
+    }
+    out[['vcov.type']] <- fvcov_type
   }
   row.names(out) <- NULL
   return(out)
@@ -53,7 +64,8 @@ glance_custom_internal.lm_robust <- function(x, vcov_type = NULL, ...) {
     out <- data.frame(row.names = "firstrow")
     if (is.null(vcov_type) || !vcov_type %in% c("vector", "matrix", "function")) {
         if (x$clustered) {
-            out[['vcov.type']] <- paste0("Clustered (", x$call$clusters, ")")
+            # out[['vcov.type']] <- paste("by:", x$call$clusters)
+            out[['se_type']] <- paste("by:", x$call$clusters)
         }
     }
     row.names(out) <- NULL
@@ -71,7 +83,7 @@ glance_custom_internal.felm <- function(x, vcov_type = NULL, ...) {
     if (is.null(vcov_type) || !vcov_type %in% c("vector", "matrix", "function")) {
         if (!is.null(x$clustervar)) {
             cluster_vars = paste(names(x$clustervar), collapse = " & ")
-            out[['vcov.type']] <- paste0("Clustered (", cluster_vars, ")")
+            out[['vcov.type']] <- paste("by:", cluster_vars)
         }
     }
     row.names(out) <- NULL
@@ -90,7 +102,7 @@ glance_custom_internal.MP <- function(x, vcov_type = NULL, ...) {
       } else {
         cluster_vars = x$DIDparams$idname
       }
-      out[['vcov.type']] <- paste0("Clustered (", cluster_vars, ")")
+      out[['vcov.type']] <- paste("by:", cluster_vars)
     }
   }
   row.names(out) <- NULL