+++ NEWS.md
@@ -4,7 +4,7 @@
 
 * Improved `vcov` argument handling for `fixest` models (#357 by @grantmcdermott)
 * Fix display of `fixest::i()` variables and interactions (#361 by @grantmcdermott)
-* Consistent display of clustered SEs (#356 and #363 by @grantmcdermott)
+* Consistent display of clustered SEs (#356, #363 and #366 by @grantmcdermott)
 
 `datasummary_correlation`: